draft v0.2
The cryptographic details section has been reordered so that credential
validation is discussed before the over- and double-spending prevention.

The introduction has been rewritten and rephramed as an abstract (but
also includes in-line citations, so this should be split back up)

Motivation section has been removed, and will be reintroduced later.

The overview section has been modified for clarity and correctness,
mainly with regards to the different user roles.

Co-authored-by: nopara73 <adam.ficsor73@gmail.com>
Co-authored-by: Seres István András <seresistvanandras@gmail.com>