Improve Bitcoin privacy related citations
Malte Möser's citation was replaced with the updated version with
co-authors.

RH13 seems to be the first academic work that studied the Bitcoin
transaction privacy, although published in 2013 it dates back to 2011
and the e-print was cited in 2012.

I was not familiar with OKH13 or AKRSC13 before today, but both
comparable to the others we already cite which discuss heuristical
attribution of transaction graph components.

These should be ordered by precedence to the best of my ability (going
by IACR upload date for RS13 and AKRSC13).