Eliminate serial number attribute
This is the serial number optimization suggested by Ruben Somsen (#42)

Instead of `C_{s_i}` being a commitment to the serial number, instead
the user provide `S_i = G_s^{r_i}` using the randomness in the amount
attribute, with a corresponding DLEQ. Since `r_i` is comitted to by the
MAC and this group element is not randomized by `z` it can be used as a
serial number.

This optimization eliminates one group element per credential request
and one field element per presentation, and requires a simpler proof
when pi^MAC and pi^serial are combined.

Also moves Credential Issuance to be after Credential Requests, so to
make it easier to understand Credential Presentation.

Closes #42

Co-authored-by: RubenSomsen <rsomsen@gmail.com>