Document WabiSabi based protocol
This document describes a concrete but general protocol which uses
WabiSabi credentials to collaboratively create transactions.

The specifics of ransaction structure are out of scope.

The SVG files are generated with `plantuml -tsvg`, and should arguably
not be in the repository, but github will not render plantuml source
files directly.

See also #64

Co-authored-by: nopara73 <adam.ficsor73@gmail.com>